sprint one compilation one, Anna-Dunya-Zach, Auth-Post/Comment-Models/Debugging/Votes
