+++ project/views.py
@@ -14,42 +14,6 @@ def index(request):
 
     return render(request, template_name, {"posts": posts})
 
-def signup_view(request):
-    if request.method == 'POST':
-        form = SignUpForm(request.POST)
-        if form.is_valid():
-            data= form.cleaned_data
-            user = Profile.objects.create_user(
-                username=data['username'],
-                password=data['password'],
-                display_name=data['display_name'],
-            )
-            login(request, user)
-            return HttpResponseRedirect(reverse('home'))
-    else:
-        form = SignUpForm
-    return render(request, 'generic_form.html', {'form': form})
-    
-def login_view(request):
-    if request.method == "POST":
-        form = LoginForm(request.POST)
-        if form.is_valid():
-            data = form.cleaned_data
-            user = authenticate(
-                request, username=data["username"], password=data["password"]
-            )
-            if user:
-                login(request, user)
-                return HttpResponseRedirect(
-                    request.GET.get("next", reverse("homepage"))
-                )
-    else:
-        form = LoginForm()
-    return render(request, "generic_form.html", {"form": form})
-
-def logout_view(request):
-    logout(request)
-    return HttpResponseRedirect(reverse("homepage"))
 
 def edit(request, id):
     if not request.user.is_staff or request.user.author == Post.post_creator:
